Add msg type + tests
